refactor(api): use scenario's blm in run service
