+++ api/apps/api/src/modules/scenarios/marxan-run/index.ts
@@ -1,2 +1,4 @@
-export { RunService, notFound, NotFound } from './run.service';
+export { RunService } from './run.service';
 export { MarxanRunModule } from './marxan-run.module';
+export { NotFound } from './cancel.handler';
+export { notFound } from './cancel.handler';
