+++ api/apps/api/src/modules/scenarios/marxan-run/run.service.spec.ts
@@ -9,14 +9,12 @@ import { ProgressData } from '@marxan/scenario-run-queue';
 import { ApiEventsService } from '@marxan-api/modules/api-events/api-events.service';
 import { CreateApiEventDTO } from '@marxan-api/modules/api-events/dto/create.api-event.dto';
 import { Scenario } from '../scenario.api.entity';
-import {
-  blmDefaultToken,
-  notFound,
-  runEventsToken,
-  runQueueToken,
-  RunService,
-} from './run.service';
+import { RunService } from './run.service';
 import { AssetsService } from './assets.service';
+import { blmDefaultToken, runEventsToken, runQueueToken } from './tokens';
+import { RunHandler } from './run.handler';
+import { CancelHandler, notFound } from './cancel.handler';
+import { EventsHandler } from './events.handler';
 
 let fixtures: PromiseType<ReturnType<typeof getFixtures>>;
 let runService: RunService;
@@ -32,7 +30,7 @@ test(`scheduling job`, async () => {
   fixtures.GivenAssetsAvailable();
 
   // when
-  await runService.run('scenario-1');
+  await runService.run({ id: 'scenario-1' });
 
   // then
   fixtures.ThenShouldUpdateScenario();
@@ -41,19 +39,34 @@ test(`scheduling job`, async () => {
   fixtures.ThenShouldUseDefaultBlm();
 });
 
-test(`scheduling job with blm`, async () => {
-  fixtures.setupMocksForSchedulingJobs();
+test(`scheduling job with overriding blm`, async () => {
+  fixtures.setupMocksForSchedulingJobs(() => `1234`);
   // given
   fixtures.GivenAssetsAvailable();
 
   // when
-  await runService.run('scenario-1', -123);
+  await runService.run({ id: 'scenario-1', boundaryLengthModifier: 78 }, -123);
 
   // then
   fixtures.ThenShouldUpdateScenario();
-  fixtures.ThenShouldEmitSubmittedEvent();
+  fixtures.ThenShouldEmitSubmittedEvent(`1234`);
   fixtures.ThenShouldAddJob();
   fixtures.ThenShouldUseBlm(-123);
+});
+
+test(`scheduling job with scenario that has blm`, async () => {
+  fixtures.setupMocksForSchedulingJobs(() => `1234`);
+  // given
+  fixtures.GivenAssetsAvailable();
+
+  // when
+  await runService.run({ id: 'scenario-1', boundaryLengthModifier: 78 });
+
+  // then
+  fixtures.ThenShouldUpdateScenario();
+  fixtures.ThenShouldEmitSubmittedEvent(`1234`);
+  fixtures.ThenShouldAddJob();
+  fixtures.ThenShouldUseBlm(78);
 });
 
 test(`scheduling job for scenario without assets`, async () => {
@@ -62,7 +75,7 @@ test(`scheduling job for scenario without assets`, async () => {
   fixtures.GivenAssetsNotAvailable();
 
   // when
-  const result = runService.run('scenario-1');
+  const result = runService.run({ id: 'scenario-1' });
 
   // then
   await expect(result).rejects.toBeDefined();
@@ -174,6 +187,9 @@ async function getFixtures() {
   };
   const testingModule = await Test.createTestingModule({
     providers: [
+      RunHandler,
+      CancelHandler,
+      EventsHandler,
       {
         provide: blmDefaultToken,
         useValue: 42,
