+++ api/apps/api/src/modules/scenarios/marxan-run/tokens.ts
@@ -1,0 +1,3 @@
+export const blmDefaultToken = Symbol('blm default token');
+export const runQueueToken = Symbol('run queue token');
+export const runEventsToken = Symbol('run events token');
